Some documentation to explain what the code does
This document explains the coding for a Youtube videos program written in Angular 5 and attempting to use Angular universal SSR
The client side code runs find with the ng serve -p=1234 or which ever port you wish to use, with the videos running in an iframe.  
However my attempt at running this via Node.js with Node dist/server.js has a bug which I'm trying to fix.